# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * 내신 계산기 & 특목고 입시 컨설팅 API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * @nullable
 */
export type RegionalPolicyGradingScale = { [key: string]: unknown } | null;

/**
 * @nullable
 */
export type RegionalPolicySemesterWeights = { [key: string]: unknown } | null;

export interface RegionalPolicy {
  id: number;
  regionName: string;
  targetYear: number;
  totalMaxScore: number;
  academicScore: number;
  attendanceScore: number;
  volunteerScore: number;
  activityScore: number;
  /** @nullable */
  gradingScale?: RegionalPolicyGradingScale;
  /** @nullable */
  semesterWeights?: RegionalPolicySemesterWeights;
  createdAt?: string;
}

/**
 * @nullable
 */
export type RegionalPolicyInputGradingScale = { [key: string]: unknown } | null;

/**
 * @nullable
 */
export type RegionalPolicyInputSemesterWeights = { [key: string]: unknown } | null;

export interface RegionalPolicyInput {
  regionName: string;
  targetYear: number;
  totalMaxScore: number;
  academicScore: number;
  attendanceScore: number;
  volunteerScore: number;
  activityScore: number;
  /** @nullable */
  gradingScale?: RegionalPolicyInputGradingScale;
  /** @nullable */
  semesterWeights?: RegionalPolicyInputSemesterWeights;
}

export interface Grade {
  id: number;
  userId: string;
  semester: string;
  subject: string;
  /** @nullable */
  rawScore?: number | null;
  /** @nullable */
  standardDeviation?: number | null;
  /** @nullable */
  average?: number | null;
  /** @nullable */
  rank?: number | null;
  /** @nullable */
  topPercentage?: number | null;
  /** @nullable */
  achievementLevel?: string | null;
  /** @nullable */
  totalStudents?: number | null;
  createdAt?: string;
}

export interface GradeInput {
  semester: string;
  subject: string;
  /** @nullable */
  rawScore?: number | null;
  /** @nullable */
  standardDeviation?: number | null;
  /** @nullable */
  average?: number | null;
  /** @nullable */
  rank?: number | null;
  /** @nullable */
  topPercentage?: number | null;
  /** @nullable */
  achievementLevel?: string | null;
  /** @nullable */
  totalStudents?: number | null;
}

export interface GradeBulk {
  grades: GradeInput[];
}

export interface Activity {
  id: number;
  userId: string;
  semester: string;
  activityType: string;
  /** @nullable */
  details?: string | null;
  /** @nullable */
  points?: number | null;
  createdAt?: string;
}

export interface ActivityInput {
  semester: string;
  activityType: string;
  /** @nullable */
  details?: string | null;
  /** @nullable */
  points?: number | null;
}

export interface StudentRecord {
  id: number;
  userId: string;
  subject: string;
  /** @nullable */
  recordText?: string | null;
  createdAt?: string;
}

export interface StudentRecordInput {
  subject: string;
  /** @nullable */
  recordText?: string | null;
}

export interface StudentProfile {
  userId: string;
  /** @nullable */
  regionId?: number | null;
  targetSchoolTypes?: string[];
  updatedAt?: string;
}

export interface StudentProfileInput {
  /** @nullable */
  regionId?: number | null;
  targetSchoolTypes?: string[];
}

export type ScoreBreakdownSemesterScores = { [key: string]: unknown };

export interface ScoreBreakdown {
  academicScore?: number;
  attendanceScore?: number;
  volunteerScore?: number;
  activityScore?: number;
  semesterScores?: ScoreBreakdownSemesterScores;
}

export interface SchoolResult {
  schoolName: string;
  schoolType: string;
  /** @nullable */
  estimatedCutoff?: number | null;
  status: string;
  achievementRate?: number;
}

export interface GapItem {
  area: string;
  description: string;
  deficit: number;
}

export interface CalculationResult {
  totalScore: number;
  maxPossibleScore?: number;
  achievementRate?: number;
  breakdown: ScoreBreakdown;
  schoolResults: SchoolResult[];
  gapAnalysis?: GapItem[];
}

export interface ConsultingRequest {
  targetSchoolType: string;
}

export interface ConsultingResult {
  academicExcellence?: string;
  inquiryCapability?: string;
  weaknessesAndPlan?: string;
  analysis: string;
}