# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: 내신 계산기 & 특목고 입시 컨설팅 API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: policies
    description: 지역별 입시 정책
  - name: grades
    description: 성적 관리
  - name: activities
    description: 활동 내역
  - name: records
    description: 학생 생기부
  - name: consulting
    description: AI 컨설팅
  - name: results
    description: 점수 계산 결과

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /policies:
    get:
      operationId: listPolicies
      tags: [policies]
      summary: 지역별 정책 목록 조회
      responses:
        "200":
          description: 정책 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RegionalPolicy"
    post:
      operationId: createPolicy
      tags: [policies]
      summary: 정책 생성 (관리자)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegionalPolicyInput"
      responses:
        "201":
          description: 생성된 정책
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegionalPolicy"

  /policies/{id}:
    get:
      operationId: getPolicy
      tags: [policies]
      summary: 정책 단건 조회
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 정책
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegionalPolicy"
    put:
      operationId: updatePolicy
      tags: [policies]
      summary: 정책 수정 (관리자)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegionalPolicyInput"
      responses:
        "200":
          description: 수정된 정책
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegionalPolicy"
    delete:
      operationId: deletePolicy
      tags: [policies]
      summary: 정책 삭제 (관리자)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: 삭제 완료

  /grades:
    get:
      operationId: listGrades
      tags: [grades]
      summary: 내 성적 목록 조회
      responses:
        "200":
          description: 성적 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Grade"
    post:
      operationId: createGrade
      tags: [grades]
      summary: 성적 입력
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GradeInput"
      responses:
        "201":
          description: 생성된 성적
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Grade"

  /grades/{id}:
    put:
      operationId: updateGrade
      tags: [grades]
      summary: 성적 수정
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GradeInput"
      responses:
        "200":
          description: 수정된 성적
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Grade"
    delete:
      operationId: deleteGrade
      tags: [grades]
      summary: 성적 삭제
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: 삭제 완료

  /grades/bulk:
    post:
      operationId: bulkSaveGrades
      tags: [grades]
      summary: 성적 일괄 저장 (upsert)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GradeBulk"
      responses:
        "200":
          description: 저장된 성적 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Grade"

  /activities:
    get:
      operationId: listActivities
      tags: [activities]
      summary: 내 활동 목록 조회
      responses:
        "200":
          description: 활동 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Activity"
    post:
      operationId: createActivity
      tags: [activities]
      summary: 활동 입력
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivityInput"
      responses:
        "201":
          description: 생성된 활동
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"

  /activities/{id}:
    put:
      operationId: updateActivity
      tags: [activities]
      summary: 활동 수정
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivityInput"
      responses:
        "200":
          description: 수정된 활동
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
    delete:
      operationId: deleteActivity
      tags: [activities]
      summary: 활동 삭제
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: 삭제 완료

  /records:
    get:
      operationId: listRecords
      tags: [records]
      summary: 내 생기부 목록 조회
      responses:
        "200":
          description: 생기부 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudentRecord"
    post:
      operationId: createRecord
      tags: [records]
      summary: 생기부 입력
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentRecordInput"
      responses:
        "201":
          description: 생성된 생기부
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentRecord"

  /records/{id}:
    put:
      operationId: updateRecord
      tags: [records]
      summary: 생기부 수정
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentRecordInput"
      responses:
        "200":
          description: 수정된 생기부
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentRecord"
    delete:
      operationId: deleteRecord
      tags: [records]
      summary: 생기부 삭제
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: 삭제 완료

  /student-profile:
    get:
      operationId: getStudentProfile
      tags: [grades]
      summary: 학생 프로필 (지역, 목표학교) 조회
      responses:
        "200":
          description: 학생 프로필
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProfile"
    put:
      operationId: saveStudentProfile
      tags: [grades]
      summary: 학생 프로필 저장
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentProfileInput"
      responses:
        "200":
          description: 저장된 프로필
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProfile"

  /results/calculate:
    get:
      operationId: calculateResults
      tags: [results]
      summary: 점수 계산 및 결과 조회
      responses:
        "200":
          description: 계산 결과
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalculationResult"

  /consulting/analyze:
    post:
      operationId: analyzeConsulting
      tags: [consulting]
      summary: AI 입시 컨설팅 분석
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsultingRequest"
      responses:
        "200":
          description: AI 분석 결과
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsultingResult"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    RegionalPolicy:
      type: object
      required: [id, regionName, targetYear, totalMaxScore, academicScore, attendanceScore, volunteerScore, activityScore]
      properties:
        id:
          type: integer
        regionName:
          type: string
        targetYear:
          type: integer
        totalMaxScore:
          type: number
        academicScore:
          type: number
        attendanceScore:
          type: number
        volunteerScore:
          type: number
        activityScore:
          type: number
        gradingScale:
          type: ["object", "null"]
        semesterWeights:
          type: ["object", "null"]
        createdAt:
          type: string

    RegionalPolicyInput:
      type: object
      required: [regionName, targetYear, totalMaxScore, academicScore, attendanceScore, volunteerScore, activityScore]
      properties:
        regionName:
          type: string
        targetYear:
          type: integer
        totalMaxScore:
          type: number
        academicScore:
          type: number
        attendanceScore:
          type: number
        volunteerScore:
          type: number
        activityScore:
          type: number
        gradingScale:
          type: ["object", "null"]
        semesterWeights:
          type: ["object", "null"]

    Grade:
      type: object
      required: [id, userId, semester, subject]
      properties:
        id:
          type: integer
        userId:
          type: string
        semester:
          type: string
        subject:
          type: string
        rawScore:
          type: ["number", "null"]
        standardDeviation:
          type: ["number", "null"]
        average:
          type: ["number", "null"]
        rank:
          type: ["integer", "null"]
        topPercentage:
          type: ["number", "null"]
        achievementLevel:
          type: ["string", "null"]
        totalStudents:
          type: ["integer", "null"]
        createdAt:
          type: string

    GradeInput:
      type: object
      required: [semester, subject]
      properties:
        semester:
          type: string
        subject:
          type: string
        rawScore:
          type: ["number", "null"]
        standardDeviation:
          type: ["number", "null"]
        average:
          type: ["number", "null"]
        rank:
          type: ["integer", "null"]
        topPercentage:
          type: ["number", "null"]
        achievementLevel:
          type: ["string", "null"]
        totalStudents:
          type: ["integer", "null"]

    GradeBulk:
      type: object
      required: [grades]
      properties:
        grades:
          type: array
          items:
            $ref: "#/components/schemas/GradeInput"

    Activity:
      type: object
      required: [id, userId, semester, activityType]
      properties:
        id:
          type: integer
        userId:
          type: string
        semester:
          type: string
        activityType:
          type: string
        details:
          type: ["string", "null"]
        points:
          type: ["number", "null"]
        createdAt:
          type: string

    ActivityInput:
      type: object
      required: [semester, activityType]
      properties:
        semester:
          type: string
        activityType:
          type: string
        details:
          type: ["string", "null"]
        points:
          type: ["number", "null"]

    StudentRecord:
      type: object
      required: [id, userId, subject]
      properties:
        id:
          type: integer
        userId:
          type: string
        subject:
          type: string
        recordText:
          type: ["string", "null"]
        createdAt:
          type: string

    StudentRecordInput:
      type: object
      required: [subject]
      properties:
        subject:
          type: string
        recordText:
          type: ["string", "null"]

    StudentProfile:
      type: object
      required: [userId]
      properties:
        userId:
          type: string
        regionId:
          type: ["integer", "null"]
        targetSchoolTypes:
          type: array
          items:
            type: string
        updatedAt:
          type: string

    StudentProfileInput:
      type: object
      properties:
        regionId:
          type: ["integer", "null"]
        targetSchoolTypes:
          type: array
          items:
            type: string

    CalculationResult:
      type: object
      required: [totalScore, breakdown, schoolResults]
      properties:
        totalScore:
          type: number
        maxPossibleScore:
          type: number
        achievementRate:
          type: number
        breakdown:
          $ref: "#/components/schemas/ScoreBreakdown"
        schoolResults:
          type: array
          items:
            $ref: "#/components/schemas/SchoolResult"
        gapAnalysis:
          type: array
          items:
            $ref: "#/components/schemas/GapItem"

    ScoreBreakdown:
      type: object
      properties:
        academicScore:
          type: number
        attendanceScore:
          type: number
        volunteerScore:
          type: number
        activityScore:
          type: number
        semesterScores:
          type: object

    SchoolResult:
      type: object
      required: [schoolName, schoolType, status]
      properties:
        schoolName:
          type: string
        schoolType:
          type: string
        estimatedCutoff:
          type: ["number", "null"]
        status:
          type: string
        achievementRate:
          type: number

    GapItem:
      type: object
      required: [area, description, deficit]
      properties:
        area:
          type: string
        description:
          type: string
        deficit:
          type: number

    ConsultingRequest:
      type: object
      required: [targetSchoolType]
      properties:
        targetSchoolType:
          type: string

    ConsultingResult:
      type: object
      required: [analysis]
      properties:
        academicExcellence:
          type: string
        inquiryCapability:
          type: string
        weaknessesAndPlan:
          type: string
        analysis:
          type: string