# lib/api-zod/src/generated/types/activityInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * 내신 계산기 & 특목고 입시 컨설팅 API
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityInput {
  semester: string;
  activityType: string;
  /** @nullable */
  details?: string | null;
  /** @nullable */
  points?: number | null;
}